+++ NEWS.md
@@ -1,2 +1,5 @@
+# r_utils 0.0.3 -- 2025-12-27
+* implement a convenient function to report database structure
+
 # r_utils 0.0.2 -- 2025-12-27
 * implement a convenient function to gather all defined functions, tests in one file and describe the project file structure